+++ lib/src/templates/template_constants.dart
@@ -7,6 +7,8 @@ const String kTemplateNameService = 'service';
 const String kTemplateNameView = 'view';
 const String kTemplateNameWidget = 'widget';
 const String kTemplateNameAppWrite = 'appwrite';
+const String kTemplateNameFirebase = 'firebase';
+const String kTemplateNameSupabase = 'supabase';
 
 // ------- Template Types --------
 const String kTemplateTypeEmpty = 'empty';